Added answer outputting, added support of floating point answers

+++ app/src/main/java/com/example/smartcalculator/Solution/PartCalculator.java
@@ -13,11 +13,11 @@ public class PartCalculator {
     public static void main(String[] args)
     {
         Coefficients map1 = new Coefficients();
-        map1.put(0,1);
-        map1.put(1,1);
+        map1.put(0,1.0);
+        map1.put(1,1.0);
         Coefficients map2 = new Coefficients();
-        map2.put(0,1);
-        map2.put(1,1);
+        map2.put(0,1.0);
+        map2.put(1,1.0);
         Coefficients sum = addTwoTrees(map1, map2);
         Coefficients mult = multiplyTwoTrees(map1, map2);
     }
@@ -26,7 +26,7 @@ public class PartCalculator {
     public static Coefficients subtractTwoTrees(Coefficients map1, Coefficients map2)
     {
         Coefficients result = new Coefficients(map1);
-        for (Map.Entry<Integer, Integer> entry : map2.entrySet())
+        for (Map.Entry<Integer, Double> entry : map2.entrySet())
         {
             if (result.containsKey(entry.getKey()))
             {
@@ -44,7 +44,7 @@ public class PartCalculator {
     public static Coefficients addTwoTrees(Coefficients map1, Coefficients map2)
     {
         Coefficients result = new Coefficients(map1);
-        for (Map.Entry<Integer, Integer> entry : map2.entrySet())
+        for (Map.Entry<Integer, Double> entry : map2.entrySet())
         {
             if (result.containsKey(entry.getKey()))
             {
@@ -62,9 +62,9 @@ public class PartCalculator {
     public static Coefficients multiplyTwoTrees(Coefficients map1, Coefficients map2)
     {
         Coefficients result = new Coefficients();
-        for (Map.Entry<Integer, Integer> entry1 : map1.entrySet())
+        for (Map.Entry<Integer, Double> entry1 : map1.entrySet())
         {
-            for (Map.Entry<Integer, Integer> entry2 : map2.entrySet())
+            for (Map.Entry<Integer, Double> entry2 : map2.entrySet())
             {
                 int power = entry1.getKey() + entry2.getKey();
                 if (result.containsKey(power))
